+++ setup.py
@@ -37,8 +37,8 @@ setup(
             'calibration_node = ros2_vive_controller.calibration_node:main',
             # The Teleop Bridge (Filename: teleop_bridge_node.py)
             'teleop_bridge_node = ros2_vive_controller.teleop_bridge_node:main',
-            'tracker_node = ros2_vive_controller.tracker_node:main',
-            'bimanual_tracker_node = ros2_vive_controller.bimanual_tracker_node:main',
+            'tracker_3_0_node = ros2_vive_controller.tracker_3_0_node:main',
+            'bimanual_tracker_3_0_node = ros2_vive_controller.bimanual_tracker_3_0_node:main',
         ],
     },
 )
